+++ pyvela/pyvela/pyvela_script.py
@@ -12,17 +12,18 @@ import emcee
 import numpy as np
 import pint
 
+import pint.models
 import pyvela
 from pyvela import SPNTA
 from pyvela import Vela as vl
 
 
 def info_dict(args):
-    return {
+    info_dict = {
         "input": {
-            "par_file": args.par_file,
-            "tim_file": args.tim_file,
-            "prior_file": args.prior_file,
+            "par_file": os.path.basename(args.par_file),
+            "tim_file": os.path.basename(args.tim_file),
+            "prior_file": os.path.basename(args.prior_file),
             "cheat_prior_scale": args.cheat_prior_scale,
         },
         "sampler": {
@@ -44,20 +45,74 @@ def info_dict(args):
         },
     }
 
+    if args.truth is not None:
+        info_dict["input"]["truth_par_file"] = os.path.basename(args.truth)
+
+    return info_dict
+
 
 def parse_args(argv):
     parser = ArgumentParser(
         prog="pyvela",
-        description="A command line interface for the Vela.jl pulsar timing & noise analysis package",
+        description="A command line interface for the Vela.jl pulsar timing &"
+        " noise analysis package. Uses emcee for sampling. This may not be "
+        "appropriate for more complex datasets. Write your own scripts for "
+        "such cases.",
     )
-    parser.add_argument("par_file")
-    parser.add_argument("tim_file")
-    parser.add_argument("-P", "--prior_file")
-    parser.add_argument("--cheat_prior_scale", default=10, type=float)
-    parser.add_argument("-o", "--outdir", default="pyvela_results")
-    parser.add_argument("-N", "--nsteps", default=6000, type=int)
-    parser.add_argument("-b", "--burnin", default=1500, type=int)
-    parser.add_argument("-t", "--thin", default=100, type=int)
+    parser.add_argument(
+        "par_file",
+        help="The pulsar ephemeris file. Should be readable using PINT. The "
+        "uncertainties listed in the file will be used for 'cheat' priors where applicable.",
+    )
+    parser.add_argument(
+        "tim_file", help="The pulsar TOA file. Should be readable using PINT."
+    )
+    parser.add_argument(
+        "-P",
+        "--prior_file",
+        help="A JSON file containing the prior distributions for each free parameter.",
+    )
+    parser.add_argument(
+        "-T",
+        "--truth",
+        help="Pulsar ephemeris file containing the true timing and noise parameter values. "
+        "Relevant for simulation studies.",
+    )
+    parser.add_argument(
+        "-C",
+        "--cheat_prior_scale",
+        default=50,
+        type=float,
+        help="The scale factor by which the frequentist uncertainties are multiplied to "
+        "get the 'cheat' prior distributions.",
+    )
+    parser.add_argument(
+        "-o",
+        "--outdir",
+        default="pyvela_results",
+        help="The output directory. Will throw an error if it already exists.",
+    )
+    parser.add_argument(
+        "-N",
+        "--nsteps",
+        default=6000,
+        type=int,
+        help="Number of ensemble MCMC iterations",
+    )
+    parser.add_argument(
+        "-b",
+        "--burnin",
+        default=1500,
+        type=int,
+        help="Burn-in length for MCMC chains",
+    )
+    parser.add_argument(
+        "-t",
+        "--thin",
+        default=100,
+        type=int,
+        help="Thinning factor for MCMC chains",
+    )
 
     return parser.parse_args(argv)
 
@@ -65,8 +120,9 @@ def parse_args(argv):
 def prepare_outdir(args):
     summary_info = info_dict(args)
 
-    if not os.path.isdir(args.outdir):
-        os.mkdir(args.outdir)
+    assert not os.path.isdir(args.outdir), "The output directory already exists!"
+
+    os.mkdir(args.outdir)
     shutil.copy(args.par_file, args.outdir)
     shutil.copy(args.tim_file, args.outdir)
     if args.prior_file is not None:
@@ -74,12 +130,31 @@ def prepare_outdir(args):
     with open(f"{args.outdir}/summary.json", "w") as summary_file:
         json.dump(summary_info, summary_file, indent=4)
 
+    if args.truth is not None:
+        shutil.copy(args.truth, args.outdir)
+
+
+def get_true_values(spnta: SPNTA, args):
+    true_model = pint.models.get_model(args.truth)
+    return (
+        np.array(
+            [
+                true_model[par].value if par in true_model else np.nan
+                for par in spnta.param_names
+            ]
+        )
+        * spnta.scale_factors
+    )
+
 
 def save_spnta_attrs(spnta: SPNTA, args):
     np.savetxt(f"{args.outdir}/param_names.txt", spnta.param_names, fmt="%s")
     np.savetxt(f"{args.outdir}/param_prefixes.txt", spnta.param_prefixes, fmt="%s")
     np.savetxt(f"{args.outdir}/param_units.txt", spnta.param_units, fmt="%s")
-    np.savetxt(f"{args.outdir}/param_scale_factors.txt", spnta.scale_factors, fmt="%s")
+    np.savetxt(f"{args.outdir}/param_scale_factors.txt", spnta.scale_factors)
+
+    if args.truth is not None:
+        np.savetxt(f"{args.outdir}/param_true_values.txt", get_true_values(spnta, args))
 
 
 def save_new_parfile(